Split big sold.cc
