+++ sold.cc
@@ -23,413 +23,10 @@
 #include <string>
 #include <vector>
 
+#include "elf_binary.h"
+#include "hash.h"
 #include "ldsoconf.h"
-
-#define Elf_Ehdr Elf64_Ehdr
-#define Elf_Phdr Elf64_Phdr
-#define Elf_Dyn Elf64_Dyn
-#define Elf_Addr Elf64_Addr
-#define Elf_Sym Elf64_Sym
-#define ELF_ST_BIND(val) ELF64_ST_BIND(val)
-#define ELF_ST_TYPE(val) ELF64_ST_TYPE(val)
-#define ELF_ST_INFO(bind, type) ELF64_ST_INFO(bind, type)
-#define Elf_Rel Elf64_Rela
-#define ELF_R_SYM(val) ELF64_R_SYM(val)
-#define ELF_R_TYPE(val) ELF64_R_TYPE(val)
-#define ELF_R_INFO(sym, type) ELF64_R_INFO(sym, type)
-
-#define CHECK(r)             \
-    do {                     \
-        if (!(r)) assert(r); \
-    } while (0)
-
-namespace {
-
-bool FLAGS_LOG{true};
-
-#ifdef NOLOG
-#define LOGF(...) \
-    if (0) fprintf(stderr, __VA_ARGS__)
-#else
-#define LOGF(...) \
-    if (FLAGS_LOG) fprintf(stderr, __VA_ARGS__)
-#endif
-
-std::vector<std::string> SplitString(const std::string& str, const std::string& sep) {
-    std::vector<std::string> ret;
-    if (str.empty()) return ret;
-    size_t index = 0;
-    while (true) {
-        size_t next = str.find(sep, index);
-        ret.push_back(str.substr(index, next - index));
-        if (next == std::string::npos) break;
-        index = next + 1;
-    }
-    return ret;
-}
-
-bool HasPrefix(const std::string& str, const std::string& prefix) {
-    ssize_t size_diff = str.size() - prefix.size();
-    return size_diff >= 0 && str.substr(0, prefix.size()) == prefix;
-}
-
-uintptr_t AlignNext(uintptr_t a, uintptr_t mask = 4095) {
-    return (a + mask) & ~mask;
-}
-
-struct Range {
-    uintptr_t start;
-    uintptr_t end;
-    ptrdiff_t size() const { return end - start; }
-    Range operator+(uintptr_t offset) const { return Range{start + offset, end + offset}; }
-};
-
-struct Elf_GnuHash {
-    uint32_t nbuckets{0};
-    uint32_t symndx{0};
-    uint32_t maskwords{0};
-    uint32_t shift2{0};
-    uint8_t tail[1];
-
-    Elf_Addr* bloom_filter() { return reinterpret_cast<Elf_Addr*>(tail); }
-
-    uint32_t* buckets() { return reinterpret_cast<uint32_t*>(&bloom_filter()[maskwords]); }
-
-    uint32_t* hashvals() { return reinterpret_cast<uint32_t*>(&buckets()[nbuckets]); }
-};
-
-uint32_t CalcGnuHash(const std::string& name) {
-    uint32_t h = 5381;
-    for (unsigned char c : name) {
-        h = h * 33 + c;
-    }
-    return h;
-}
-
-uint32_t CalcHash(const std::string& name) {
-    uint32_t h = 0, g;
-    for (unsigned char c : name) {
-        h = (h << 4) + c;
-        g = h & 0xf0000000;
-        h ^= g >> 24;
-        h ^= g;
-    }
-    return h;
-}
-
-struct Elf_Hash {
-    uint32_t nbuckets{0};
-    uint32_t nchains{0};
-    uint8_t tail[0];
-
-    const uint32_t* buckets() const { return reinterpret_cast<const uint32_t*>(tail); }
-
-    const uint32_t* chains() const { return buckets() + nchains; }
-};
-
-bool IsTLS(const Elf_Sym& sym) {
-    return ELF_ST_TYPE(sym.st_info) == STT_TLS;
-}
-
-bool IsDefined(const Elf_Sym& sym) {
-    return sym.st_value || IsTLS(sym);
-}
-
-}  // namespace
-
-class ELFBinary {
-public:
-    ELFBinary(const std::string& filename, int fd, char* head, size_t size) : filename_(filename), fd_(fd), head_(head), size_(size) {
-        ehdr_ = reinterpret_cast<Elf_Ehdr*>(head);
-
-        {
-            size_t found = filename.rfind('/');
-            if (found == std::string::npos) {
-                name_ = filename;
-            } else {
-                name_ = filename.substr(found + 1);
-            }
-        }
-
-        ParsePhdrs();
-    }
-
-    ~ELFBinary() {
-        munmap(head_, size_);
-        close(fd_);
-    }
-
-    static bool IsELF(const char* p) {
-        if (strncmp(p, ELFMAG, SELFMAG)) {
-            return false;
-        }
-        if (p[EI_CLASS] != ELFCLASS64) {
-            err(1, "non 64bit ELF isn't supported yet");
-        }
-        return true;
-    }
-
-    const std::string& filename() const { return filename_; }
-
-    const Elf_Ehdr* ehdr() const { return ehdr_; }
-    const std::vector<Elf_Phdr*> phdrs() const { return phdrs_; }
-    const std::vector<Elf_Phdr*> loads() const { return loads_; }
-    const Elf_Phdr* tls() const { return tls_; }
-
-    const std::vector<std::string>& neededs() const { return neededs_; }
-    const std::string& soname() const { return soname_; }
-    const std::string& runpath() const { return runpath_; }
-    const std::string& rpath() const { return rpath_; }
-
-    const Elf_Sym* symtab() const { return symtab_; }
-    const Elf_Rel* rel() const { return rel_; }
-    size_t num_rels() const { return num_rels_; }
-    const Elf_Rel* plt_rel() const { return plt_rel_; }
-    size_t num_plt_rels() const { return num_plt_rels_; }
-
-    const char* head() const { return head_; }
-
-    const std::string& name() const { return name_; }
-
-    uintptr_t init() const { return init_; }
-    uintptr_t fini() const { return fini_; }
-    const std::vector<uintptr_t>& init_array() const { return init_array_; }
-    const std::vector<uintptr_t>& fini_array() const { return fini_array_; }
-
-    const std::map<std::string, Elf_Sym*>& GetSymbolMap() const { return syms_; }
-
-    Range GetRange() const {
-        Range range{std::numeric_limits<uintptr_t>::max(), std::numeric_limits<uintptr_t>::min()};
-        for (Elf_Phdr* phdr : loads_) {
-            range.start = std::min(range.start, phdr->p_vaddr);
-            range.end = std::max(range.end, AlignNext(phdr->p_vaddr + phdr->p_memsz));
-        }
-        return range;
-    }
-
-    bool InTLS(uintptr_t offset) const {
-        if (tls_) {
-            return tls_->p_vaddr <= offset && offset < tls_->p_vaddr + tls_->p_memsz;
-        }
-        return false;
-    }
-
-    void ReadDynSymtab() {
-        CHECK(symtab_);
-        LOGF("Read dynsymtab of %s\n", name().c_str());
-        if (gnu_hash_) {
-            const uint32_t* buckets = gnu_hash_->buckets();
-            const uint32_t* hashvals = gnu_hash_->hashvals();
-            for (int i = 0; i < gnu_hash_->nbuckets; ++i) {
-                int n = buckets[i];
-                if (!n) continue;
-                const uint32_t* hv = &hashvals[n - gnu_hash_->symndx];
-                for (Elf_Sym* sym = &symtab_[n];; ++sym) {
-                    uint32_t h2 = *hv++;
-                    const std::string name(strtab_ + sym->st_name);
-                    // LOGF("%s@%s\n", name.c_str(), name_.c_str());
-                    // TODO(hamaji): Handle version symbols.
-                    CHECK(syms_.emplace(name, sym).second);
-                    if (h2 & 1) break;
-                }
-            }
-            for (size_t n = 0; n < gnu_hash_->symndx; ++n) {
-                Elf_Sym* sym = &symtab_[n];
-                if (sym->st_name) {
-                    const std::string name(strtab_ + sym->st_name);
-                    CHECK(syms_.emplace(name, sym).second);
-                }
-            }
-        } else {
-            CHECK(hash_);
-            const uint32_t* buckets = hash_->buckets();
-            for (size_t i = 0; i < hash_->nbuckets; ++i) {
-                int n = buckets[i];
-                Elf_Sym* sym = &symtab_[n];
-                if (sym->st_name == 0) continue;
-                const std::string name(strtab_ + sym->st_name);
-                CHECK(syms_.emplace(name, sym).second);
-            }
-        }
-    }
-
-    const char* Str(uintptr_t name) { return strtab_ + name; }
-
-    char* GetPtr(uintptr_t offset) { return head_ + OffsetFromAddr(offset); }
-
-    Elf_Phdr* FindPhdr(uint64_t type) {
-        for (Elf_Phdr* phdr : phdrs_) {
-            if (phdr->p_type == type) {
-                return phdr;
-            }
-        }
-        return nullptr;
-    }
-
-    const Elf_Phdr& GetPhdr(uint64_t type) {
-        Elf_Phdr* phdr = FindPhdr(type);
-        CHECK(phdr);
-        return *phdr;
-    }
-
-private:
-    void ParsePhdrs() {
-        for (int i = 0; i < ehdr_->e_phnum; ++i) {
-            Elf_Phdr* phdr = reinterpret_cast<Elf_Phdr*>(head_ + ehdr_->e_phoff + ehdr_->e_phentsize * i);
-            phdrs_.push_back(phdr);
-            if (phdr->p_type == PT_LOAD) {
-                loads_.push_back(phdr);
-            } else if (phdr->p_type == PT_TLS) {
-                tls_ = phdr;
-            }
-        }
-
-        for (Elf_Phdr* phdr : phdrs_) {
-            if (phdr->p_type == PT_DYNAMIC) {
-                ParseDynamic(phdr->p_offset, phdr->p_filesz);
-            }
-        }
-        CHECK(!phdrs_.empty());
-    }
-
-    void ParseDynamic(size_t off, size_t size) {
-        size_t dyn_size = sizeof(Elf_Dyn);
-        CHECK(size % dyn_size == 0);
-        std::vector<Elf_Dyn*> dyns;
-        for (size_t i = 0; i < size / dyn_size; ++i) {
-            Elf_Dyn* dyn = reinterpret_cast<Elf_Dyn*>(head_ + off + dyn_size * i);
-            dyns.push_back(dyn);
-        }
-
-        uintptr_t* init_array{0};
-        uintptr_t init_arraysz{0};
-        uintptr_t* fini_array{0};
-        uintptr_t fini_arraysz{0};
-        for (Elf_Dyn* dyn : dyns) {
-            auto get_ptr = [this, dyn]() { return GetPtr(dyn->d_un.d_ptr); };
-            if (dyn->d_tag == DT_STRTAB) {
-                strtab_ = get_ptr();
-            } else if (dyn->d_tag == DT_SYMTAB) {
-                symtab_ = reinterpret_cast<Elf_Sym*>(get_ptr());
-            } else if (dyn->d_tag == DT_GNU_HASH) {
-                gnu_hash_ = reinterpret_cast<Elf_GnuHash*>(get_ptr());
-            } else if (dyn->d_tag == DT_HASH) {
-                hash_ = reinterpret_cast<Elf_Hash*>(get_ptr());
-            } else if (dyn->d_tag == DT_RELA) {
-                rel_ = reinterpret_cast<Elf_Rel*>(get_ptr());
-            } else if (dyn->d_tag == DT_RELASZ) {
-                num_rels_ = dyn->d_un.d_val / sizeof(Elf_Rel);
-            } else if (dyn->d_tag == DT_JMPREL) {
-                plt_rel_ = reinterpret_cast<Elf_Rel*>(get_ptr());
-            } else if (dyn->d_tag == DT_PLTRELSZ) {
-                num_plt_rels_ = dyn->d_un.d_val / sizeof(Elf_Rel);
-            } else if (dyn->d_tag == DT_PLTREL) {
-                CHECK(dyn->d_un.d_val == DT_RELA);
-            } else if (dyn->d_tag == DT_PLTREL) {
-                // TODO(hamaji): Check
-            } else if (dyn->d_tag == DT_REL || dyn->d_tag == DT_RELSZ || dyn->d_tag == DT_RELENT) {
-                // TODO(hamaji): Support 32bit?
-                CHECK(false);
-            } else if (dyn->d_tag == DT_INIT_ARRAY) {
-                init_array = reinterpret_cast<uintptr_t*>(get_ptr());
-            } else if (dyn->d_tag == DT_INIT_ARRAYSZ) {
-                init_arraysz = dyn->d_un.d_val;
-            } else if (dyn->d_tag == DT_FINI_ARRAY) {
-                fini_array = reinterpret_cast<uintptr_t*>(get_ptr());
-            } else if (dyn->d_tag == DT_FINI_ARRAYSZ) {
-                fini_arraysz = dyn->d_un.d_val;
-            } else if (dyn->d_tag == DT_INIT) {
-                init_ = dyn->d_un.d_ptr;
-            } else if (dyn->d_tag == DT_FINI) {
-                fini_ = dyn->d_un.d_ptr;
-            }
-        }
-        CHECK(strtab_);
-
-        ParseFuncArray(init_array, init_arraysz, &init_array_);
-        ParseFuncArray(fini_array, fini_arraysz, &fini_array_);
-
-        for (Elf_Dyn* dyn : dyns) {
-            if (dyn->d_tag == DT_NEEDED) {
-                const char* needed = strtab_ + dyn->d_un.d_val;
-                neededs_.push_back(needed);
-            } else if (dyn->d_tag == DT_SONAME) {
-                name_ = soname_ = strtab_ + dyn->d_un.d_val;
-            } else if (dyn->d_tag == DT_RUNPATH) {
-                runpath_ = strtab_ + dyn->d_un.d_val;
-            } else if (dyn->d_tag == DT_RPATH) {
-                rpath_ = strtab_ + dyn->d_un.d_val;
-            }
-        }
-    }
-
-    void ParseFuncArray(uintptr_t* array, uintptr_t size, std::vector<uintptr_t>* out) {
-        for (size_t i = 0; i < size / sizeof(uintptr_t); ++i) {
-            out->push_back(array[i]);
-        }
-    }
-
-    Elf_Addr OffsetFromAddr(Elf_Addr addr) {
-        for (Elf_Phdr* phdr : loads_) {
-            if (phdr->p_vaddr <= addr && addr < phdr->p_vaddr + phdr->p_memsz) {
-                return addr - phdr->p_vaddr + phdr->p_offset;
-            }
-        }
-        LOGF("Address %llx cannot be resolved\n", static_cast<long long>(addr));
-        abort();
-    }
-
-    const std::string filename_;
-    int fd_;
-    char* head_;
-    size_t size_;
-
-    Elf_Ehdr* ehdr_{nullptr};
-    std::vector<Elf_Phdr*> phdrs_;
-    std::vector<Elf_Phdr*> loads_;
-    Elf_Phdr* tls_{nullptr};
-    const char* strtab_{nullptr};
-    Elf_Sym* symtab_{nullptr};
-
-    std::vector<std::string> neededs_;
-    std::string soname_;
-    std::string runpath_;
-    std::string rpath_;
-
-    Elf_Rel* rel_{nullptr};
-    size_t num_rels_{0};
-    Elf_Rel* plt_rel_{nullptr};
-    size_t num_plt_rels_{0};
-
-    Elf_GnuHash* gnu_hash_{nullptr};
-    Elf_Hash* hash_{nullptr};
-
-    uintptr_t init_{0};
-    uintptr_t fini_{0};
-    std::vector<uintptr_t> init_array_;
-    std::vector<uintptr_t> fini_array_;
-
-    std::string name_;
-    std::map<std::string, Elf_Sym*> syms_;
-};
-
-std::unique_ptr<ELFBinary> ReadELF(const std::string& filename) {
-    int fd = open(filename.c_str(), O_RDONLY);
-    if (fd < 0) err(1, "open failed: %s", filename.c_str());
-
-    size_t size = lseek(fd, 0, SEEK_END);
-    if (size < 8 + 16) err(1, "too small file: %s", filename.c_str());
-
-    size_t mapped_size = (size + 0xfff) & ~0xfff;
-
-    char* p = (char*)mmap(NULL, mapped_size, PROT_READ | PROT_WRITE, MAP_PRIVATE, fd, 0);
-    if (p == MAP_FAILED) err(1, "mmap failed: %s", filename.c_str());
-
-    if (ELFBinary::IsELF(p)) {
-        return std::make_unique<ELFBinary>(filename.c_str(), fd, p, mapped_size);
-    }
-    err(1, "unknown file format: %s", filename.c_str());
-}
+#include "utils.h"
 
 class StrtabBuilder {
 public:
